Go: Move `go` invocations to separate file
