+++ go/extractor/cli/go-autobuilder/go-autobuilder.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
 	"log"
 	"net/url"
@@ -17,6 +16,7 @@ import (
 
 	"github.com/github/codeql-go/extractor/autobuilder"
 	"github.com/github/codeql-go/extractor/diagnostics"
+	"github.com/github/codeql-go/extractor/toolchain"
 	"github.com/github/codeql-go/extractor/util"
 )
 
@@ -56,43 +56,9 @@ Build behavior:
 	fmt.Fprintf(os.Stderr, "Usage:\n\n  %s\n", os.Args[0])
 }
 
-var goVersion = ""
-
-// Returns the current Go version as returned by 'go version', e.g. go1.14.4
-func getEnvGoVersion() string {
-	if goVersion == "" {
-		// Since Go 1.21, running 'go version' in a directory with a 'go.mod' file will attempt to
-		// download the version of Go specified in there. That may either fail or result in us just
-		// being told what's already in 'go.mod'. Setting 'GOTOOLCHAIN' to 'local' will force it
-		// to use the local Go toolchain instead.
-		cmd := exec.Command("go", "version")
-		cmd.Env = append(os.Environ(), "GOTOOLCHAIN=local")
-		out, err := cmd.CombinedOutput()
-
-		if err != nil {
-			log.Fatalf("Unable to run the go command, is it installed?\nError: %s", err.Error())
-		}
-
-		goVersion = parseGoVersion(string(out))
-	}
-	return goVersion
-}
-
-// The 'go version' command may output warnings on separate lines before
-// the actual version string is printed. This function parses the output
-// to retrieve just the version string.
-func parseGoVersion(data string) string {
-	var lastLine string
-	sc := bufio.NewScanner(strings.NewReader(data))
-	for sc.Scan() {
-		lastLine = sc.Text()
-	}
-	return strings.Fields(lastLine)[2]
-}
-
 // Returns the current Go version in semver format, e.g. v1.14.4
 func getEnvGoSemVer() string {
-	goVersion := getEnvGoVersion()
+	goVersion := toolchain.GetEnvGoVersion()
 	if !strings.HasPrefix(goVersion, "go") {
 		log.Fatalf("Expected 'go version' output of the form 'go1.2.3'; got '%s'", goVersion)
 	}
@@ -780,7 +746,7 @@ func getBuildInfo(emitDiagnostics bool) BuildInfo {
 
 // Build the project and run the extractor.
 func installDependenciesAndBuild() {
-	log.Printf("Autobuilder was built with %s, environment has %s\n", runtime.Version(), getEnvGoVersion())
+	log.Printf("Autobuilder was built with %s, environment has %s\n", runtime.Version(), toolchain.GetEnvGoVersion())
 
 	srcdir := getSourceDir()
 
@@ -1110,12 +1076,6 @@ func (v versionInfo) String() string {
 		v.goModVersion, v.goModVersionFound, v.goEnvVersion, v.goEnvVersionFound)
 }
 
-// Check if Go is installed in the environment.
-func isGoInstalled() bool {
-	_, err := exec.LookPath("go")
-	return err == nil
-}
-
 // Get the version of Go to install and output it to stdout as json.
 func identifyEnvironment() {
 	var v versionInfo
@@ -1123,9 +1083,9 @@ func identifyEnvironment() {
 	goVersionInfo := tryReadGoDirective(buildInfo)
 	v.goModVersion, v.goModVersionFound = goVersionInfo.Version, goVersionInfo.Found
 
-	v.goEnvVersionFound = isGoInstalled()
+	v.goEnvVersionFound = toolchain.IsInstalled()
 	if v.goEnvVersionFound {
-		v.goEnvVersion = getEnvGoVersion()[2:]
+		v.goEnvVersion = toolchain.GetEnvGoVersion()[2:]
 	}
 
 	msg, versionToInstall := getVersionToInstall(v)
