+++ go/extractor/cli/go-autobuilder/go-autobuilder_test.go
@@ -21,19 +21,6 @@ func TestGetImportPathFromRepoURL(t *testing.T) {
 	}
 }
 
-func TestParseGoVersion(t *testing.T) {
-	tests := map[string]string{
-		"go version go1.18.9 linux/amd64": "go1.18.9",
-		"warning: GOPATH set to GOROOT (/usr/local/go) has no effect\ngo version go1.18.9 linux/amd64": "go1.18.9",
-	}
-	for input, expected := range tests {
-		actual := parseGoVersion(input)
-		if actual != expected {
-			t.Errorf("Expected parseGoVersion(\"%s\") to be \"%s\", but got \"%s\".", input, expected, actual)
-		}
-	}
-}
-
 func TestGetVersionToInstall(t *testing.T) {
 	tests := map[versionInfo]string{
 		// getVersionWhenGoModVersionNotFound()
